requirements.txt
click==8.1.3
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
PyJWT==2.6.0
Werkzeug==2.2.2
zdm-client-py==1.0.1
